requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==2.1.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
passlib==1.7.4
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.22.0
bcrypt==4.3.0
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-api-python-client==2.95.0
apscheduler==3.10.1
python-jose==3.3.0
python-dotenv>=1.0.0
setuptools>=68.0.0
wheel>=0.40.0
openpyxl
itsdangerous
pytest
httpx
flake8